testing-video/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
openai-whisper==20231117
moviepy==1.0.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
pydub==0.25.1
scipy==1.13.1
pillow==10.3.0
python-dotenv==1.0.1
